[mpim] Overridable SW prefix for ICON on LUMI
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/300

We still have users who need to build ICON on LUMI but do not have access to the
software stack directory. This MR reintroduces the ICON_SW_PREFIX environment
variable, which makes it possible to use the container and configure wrappers
as-is against a custom installation of the software stack.

Merged-by: Sergey Kosukhin <sergey.kosukhin@mpimet.mpg.de>