[mpim] OpenACC updates in ICON-Land/JSBACH
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/310

This merge contains many changes to OpenACC directives in ICON-Land/JSBACH in order
to follow the ICON OpenACC programming guidelines (e.g. putting ASYNCs everywhere).

In addition, the workaround in the aggregate routines for the cray compiler on lumi
is not necessary anymore.

Results are not changed.

See https://gitlab.dkrz.de/jsbach/jsbach/-/merge_requests/132

Merged-by: Reiner Schnur <reiner.schnur@mpimet.mpg.de>