[mpim] Remove isobaric updating within AES physics
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/327

Previously within the physics the temperature was updated assuming
an isobaric heating, and the tendencies were collected and renormalized
to be reflective of an isometric heating after physics was completed. This
was done in the routine finalize.  The logic however was convoluted, also
because some routines, such as the single moment microphsyics and the
turbulence were actually working in the isometric space, in which case
their tendencies had to be converted back and forth.  This was prone to
mistakes and not even clearly correct.  Hence this functionality has been
removed to allow a logically simpler updating.

With the changes in this MR the physics code is consistently applying the
physical tendencies at constant mass and volume in the one-moment cloud
microphysics (mig), the radiant energy transfer (rad) and the  turbulent
mixing (vdf/tmx). Changes were also made in the two-moment microphysics
although this is currently not really supported.

This changes results when running vdiff from the AES physics but a first
(short) qubicc test showed that they are similar to before. It also has
been tested on the bubble test case, and the evolution is nearly
indistinguishable.

A better implementation would call get_cvair before it is applied in
the turbulence mixing updates, so that it reflected the new moisture
state.

This commit contains new update reference and tolerance data.

Co-authored-by: Bjorn Stevens <bjorn.stevens@mpimet.mpg.de>
Approved-by: Bjorn Stevens <bjorn.stevens@mpimet.mpg.de>
Merged-by: Reiner Schnur <reiner.schnur@mpimet.mpg.de>