[mpim] Add tuning options in tmx
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/328

See the above MR for further details.

For the first production runs with tmx we have a few adjustable parameters we work with.  Some that set the global cloud liquid water, and its efficiency at scattering radiation, and hence the planetary albedo. This is important for getting a climate at about the right temperature under present day conditions. An additional one is introduced to compensate for the residual energy leak in the dynamics.

The residual dynamic energy leak is compensated by a rescaling of the dissipation of kinetic energy, through the namelist parameter `dissipation_factor` in `aes_vdf_nml` (default: 1.0).

A new stability correction to turbulent mixing has been added (Louis, 1979) that is controlled by
the namelist parameter `louis_constant_b` in `aes_vdf_nml` (default: 4.2). Setting `use_louis = .FALSE.` switches back to the previous scheme.

The value for the cloud droplet number `cloud_num` in `mo_cloud_mig` has been changed
from 200.0 (by a USE from `gscp_data`) to 50.0 . Results for vdiff therefore change
with this commit.

Co-authored-by: Junhong Lee <junhong.lee@mpimet.mpg.de>
Co-authored-by: Bjorn Stevens <bjorn.stevens@mpimet.mpg.de>
Co-authored-by: Reiner Schnur <reiner.schnur@mpimet.mpg.de>
Approved-by: Bjorn Stevens <bjorn.stevens@mpimet.mpg.de>
Merged-by: Reiner Schnur <reiner.schnur@mpimet.mpg.de>