[mpim] Include non-containerized config wrappes
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/312

For development purposes, it may be more convenient to build and run ICON
without a container. However, multiple versions of the wrappers for more or less
the same compiler versions might be confusing.

This MR suggests naming the wrappers explicitly depending on whether they are
meant to be used with containers or not; the default versions (i.e. targets of
the symbolic links 'lumi.gpu.cce'/'lumi.cpu.cce') are the containerized ones.

Co-authored-by: Sergey Kosukhin <sergey.kosukhin@mpimet.mpg.de>
Merged-by: Sergey Kosukhin <sergey.kosukhin@mpimet.mpg.de>