[mpim] Add diagnostic for 2m specific humidity in tmx
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/341

New diagnostic for 2m specific humidity in tmx, for the grid box mean
and for the tiles wtr, ice and lnd. The name is `qv2m` for the grid box
mean or `qv2m_[wtr,ice,lnd}` for the tiles.

Also added a mapping to `run/dict.iconam.mpim` so that `hus2m*` can also
be used instead of `qv2m*`.

Approved-by: Bjorn Stevens <bjorn.stevens@mpimet.mpg.de>
Merged-by: Reiner Schnur <reiner.schnur@mpimet.mpg.de>