[nwp] Cloud droplet number from MODIS climatology
See merge request icon/icon-nwp!963

The cloud droplet number from MODIS climatology is read in from external parameter file. This is done when icpl_aero_gscp is set = 3.

This is meant to be used together with the transient aerosol for ICON-CLM and ICON-Seamless.
The cdnc from MODIS is added to the extpar file by a script, similar to what is done with the LSM from ICON-O.

Test with MODIS cdnc is added to checksuite.nwp/nwpexp.run_ICON_19_R2B4_cmip_forcing

Approved-by: Axel Seifert <axel.seifert@dwd.de>
Merged-by: Daniel Rieger <daniel.rieger@dwd.de>