Urgent hotfix for RUC: emvorado ready file mechanism
See merge request icon/icon-nwp!1301

In emvorado change writing of emvorado ready files to the new ICON mechanism:
first create a .-file, write some text to the file and then move it to the final filename.

See merge requests dwd-sw/emvorado-for-icon!15 and dwd-sw/emvorado-for-icon!16

This will avoid "empty ready file" crashes of the RUC parallel routine.

Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>