Adding nblocks_e namelist option + updated logic for nproma, nblocks_c
See merge request icon/icon-nwp!1231

There are a few potential use cases for adding an `nblocks_e` namelist option to ICON. 
- For GPU runs, it could possibly be more optimal to set `nproma` based on the num_edges rather than `num_cells`, and a namelist option would help software engineers investigate this more easily
- The ICON-EXCLAIM with GT4Py dycore assumes there is only one block, which encompasses all the edges. So presently, we get the model to run by explicitly calculating `nproma` from num_edges and setting `nblocks_c=0`, which is very cumbersome for users. And we would rather think about a proper implementation in ICON and pull changes, rather than our custom implementation downstream. 

A longer-term option would be to introduce three different npromas/nblocks for cells, vertices, edges. But this is way too much work. So presently, we just want to add `nblocks_e` to ICON in a way that it touches the existing runscripts and functionality only minimally.

-------------------
So far, `nblocks_c` has been used to set `nproma` using the following logic (correct me, if I'm mistaken):

| nproma | nblocks_c |  Decision | 
| ------   | ------ | ------ | 
|  any val | >0          | Use `nblocks_c` to recompute `nproma` |
|   >0      |  <=0         | Use specified `nproma`       |

In this merge request, we want to add `nblocks_e` to the mix and tweak the logic a little so that the specification of `nproma`, `nblocks_c` and `nblocks_e` is more explicit. Only one of `nproma` or `nblocks_c` or `nblocks_e` can have a positive value, in which case the positive namelist option will be used to set the nproma eventually. If more than two options have a positive value, it is considered an ambiguous definition and the model run exits. If none of the namelist options have a positive value, then we set `nblocks_c=1` and proceed to recompute `nproma` . This is detailed in the table below.


| nproma   | nblocks_c | nblocks_e |  Decision | 
| ------   | ------    | ------    | ------ | 
|  <=0     |  <=0      |   <=0     | Set `nproma=1` |
|  <=0     |  **>0**   |   <=0     | Use `nblocks_c` to recompute `nproma` |
|  <=0     |  <=0      |   **>0**  | Use `nblocks_e` to recompute `nproma` |
|  **>0** |  <=0      |   <=0     | Use specified `nproma`       |
|  <=0     |  **>0**   |   **>0**  | Exit |
|  **>0**  |  **>0**   |  <=0      | Exit |
|  **>0**  | <=0       |   **>0**  | Exit |
|  **>0**  | **>0**    |   **>0**  | Exit |

-------------------
nproma_sub
-------------------
In this MR, we also wanted to unify the logic for `nproma_sub` and `nblocks_sub`, which currently differs a little from `nproma`, `nblocks_c`. 

Old logic

| nproma_sub | nblocks_sub |  Decision | 
| ------   | ------ | ------ | 
|  any val | >0          | Use `nblocks_c` to recompute `nproma_sub` |
|   >0      |  <=0         | Use specified `nproma_sub`       |


New logic:


|  nproma_sub | nblocks_sub | Decision |
| ------      | ------      | ------ | 
|  <=0        | **>0**      | Use `nblocks_sub` to recompute `nproma_sub` |
|   **>0**    |  <=0        | Use specified `nproma_sub`       |
|   >0        |  >0         | Throw error/exit       |
|  <=0        |  <=0        | Set `nblocks_sub=1` and recompute `nproma_sub`       |

Approved-by: Guenther Zaengl <guenther.zaengl@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>