[mpim] Add the possibility to output the ocean potential temperature in Kelvin as a new field.
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/368

A new output field 'to_k' was added that outputs the ocean potential temperature in Kelvin instead of °C.
This is necessary for GRIB output.
The conversion is done using a post_op. To this end, it was necessary to implement a new post_op POST_OP_OFFSET that adds a constant to an output field

Co-authored-by: Martin Sprengel <Martin.Sprengel@dwd.de>
Merged-by: Helmuth Haak <helmuth.haak@mpimet.mpg.de>