[fix] call interface_aes_ocean only if coupled_to_ocean
See merge request icon/icon-nwp!1317

If it is a coupled run (e.g. using `coupled_to_output = .TRUE.`) but not coupled to the ocean the call to `interface_aes_ocean` fails (as it is not constructed). This MR fixes this by calling `interface_aes_ocean` onliy if `coupled_to_ocean` is set.

Approved-by: Florian Prill <florian.prill@dwd.de>
Merged-by: Florian Prill <florian.prill@dwd.de>