[hotfix] fix alps setup
See merge request icon/icon-nwp!1329

Removed the d56 account that bb used before on alps. Now no account is specified and the default is used (s83ci for Icontest)

Added the env variable FI_CXI_SAFE_DEVMEM_COPY_THRESHOLD=0 which is needed on alps now.

Changed the regex to find the jobid, since before it would only consider numbers with 5 or more digits as jobids, but the counter reset on balfrin recently.

Deactivated cuda sanitizer for kenda test on balfrin

Approved-by: Xavier Lapillonne <xavier.lapillonne@meteoswiss.ch>
Merged-by: Daniel Hupp <daniel.hupp@meteoswiss.ch>