[nwp] exp_mch_description
See merge request icon/icon-nwp!1318

The runscripts used for testing at MeteoSwiss have a header describing Name, Purpose, and Author of the Runscript.
Since this is a comment only and the runscripts have been changed or copy+pasted for other purposes than the initial one, the information in the header were outdated on some runscripts.

I corrected the false information and renamed "Author" to "Created by" since this is a more fitting description, as most of these files were not really authored by someone (as in significant parts were written from scratch), merely taken and adapted.

Only comments were changed. This MR has no impact on the code itself.

Approved-by: Daniel Hupp <daniel.hupp@meteoswiss.ch>
Merged-by: Daniel Hupp <daniel.hupp@meteoswiss.ch>