[mpim] Fix for vdiff and tmx with Smagorinsky model
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/391

An inconsistency in the Smagorinsky model of the vdiff turbulence scheme
is fixed. So far, the TTE scheme returns turbulent diffusion coefficients
excluding the density. However, the Smagorinsky model provides mass-weighted
turbulent diffusion coefficients including the density. This inconsistency is fixed
by removing the densiy from the variables pcfm and pcfh in vdiff, which contain
the turbulent diffusion coefficients.

In tmx, the inconsistency is fixed by removing the extra multiplication with
density in the vertical diffusion solvers.

Co-authored-by: Moritz Waldmann <Moritz.Waldmann@dwd.de>  
Co-authored-by: Roland Wirth <Roland.Wirth@dwd.de>  
Co-authored-by: Kay Schaefer <kay.schaefer@uni-hamburg.de>
Approved-by: Kay Schäfer <kay.schaefer@uni-hamburg.de>
Approved-by: Moritz Waldmann <moritz.waldmann@dwd.de>
Approved-by: Roland Wirth <roland.wirth@dwd.de>
Merged-by: Reiner Schnur <reiner.schnur@mpimet.mpg.de>