[nwp] SER: fix serialization writing from GPU.
See merge request icon/icon-nwp!235

This MR corrects data management for serialize_all:
Write mode: All data is now copied to a temporary array on the
device. The data copy is then copied onto the host for writing.

Read/perturb mode: All data is read into a temporary on the host.
The temporary data is then copied into the device where it gets copied
over onto the real array.

(Compare mode: The reference data is already read in a separate array
which is subsequently copied onto the device for the comparison.)

This commit removes the `lupdate_cpu` argument, as it is not needed for
serialization anymore. If synchronization is needed, it should be done
in the code explicitly.

Approved-by: Abishek Gopal <abishek.gopal@c2sm.ethz.ch>
Merged-by: Daniel Hupp <daniel.hupp@meteoswiss.ch>