[nwp] nwpexp.run_ICON_{18,21}: remove request for VH processes for waves component
See merge request icon/icon-nwp!1360

waves does not support `proc0_shift=1` so IO PEs running on VH crash the model.
The current START_MODEL function ignores the request for VH PEs if
`proc0_shift=0`, but the generalized one we'd like to introduce with !1359 does
not. Keeping the old behavior would complicate the logic and/or introduce
unnecessary interdependencies, so this MR proposes to remove the request. The
allocation with the new function remains the same -- IO PEs are allocated on
VEs, reducing the number of worker PEs.

Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>