[mpim] ICON-Land: Flexible number of soil layers
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/399

In ICON-Land, the number of soil layers is read from the initial soil conditions
file. In previous versions, the number of soil layers was however hard coded in
some parts of the code. It is now (again) possible to run simulations with a
non-default number of soil layers without changing the source code.

Merged-by: Veronika Gayler <veronika.gayler@mpimet.mpg.de>