[nwp] Hd updates
See merge request icon/icon-nwp!1341

The URL of submodule HD is updated to the new public one:

https://gitlab.dkrz.de/hd-model/hd-couple.git

instead of the old one:

git@gitlab.dkrz.de:dwd-sw/hd-couple.git

Background story: The old Git Repo is actually a mirror of a personal git repo from the HD model author. Which was not appropriate for ICON. At the beginning, we found a solution to mirror HD code to dwd-sw. But with the ICON open source code (HD is still not a part of), license and some formalities need to be changed. Therefore, Stefan Hagemann changed the HD Git Repo to the new public one. 

After this merge, the dwd-sw/hd-couple.git will be deleted.

Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>