[nwp] Always specify model type, when creating a new varlist with vlr_add
See merge request icon/icon-nwp!1365

New variable lists are registered by calling the subroutine `vlr_add`. This subroutine has several optional arguments, with `model_type` being one of them. It is a character string that specifies the model (such as atm, oce, wave, ...), which has registered the variable list at hand. If nothing is specified, `model_type=atm` is set as a default. This turned out to be problematic for the following reasons:

1. Problems occur, if code that registers variable lists is shared by multiple models. An example of such code is `atm_dyn_iconam/mo_opt_diagnostics.f90` which is needed by `mo_pp_scheduler.f90`. It is used by the atmosphere model and the wave model. If the variable list lacks an explicit specification of the model type, the restart functionality for non-atmosphere models like `oce`, `wave`, ... breaks, as the model will include one or more variable lists for which `model_type=atm` is set. In this case, the restart module requests more than one restart file (i.e. one file for each model type).
2. The default `model_type = 'atm'` in `TYPE t_var_list` (`mo_var_list.f90`) privileges the atmosphere model, even though this code should be largely model independent. This default value was adequate in previous time when there were only the atmosphere and ocean model. Nowadays, with roughly a hand full of different models available in the ICON framework, this bias in favor of the atmosphere model should be avoided.


This merge request fixes the described restart issues, which popped up during the implementation of the restart functionality for the wave model. The issues are solved, by specifying the subroutine argument `model_type` unconditionally for all models, including the atmosphere model. In order to allow sharing of code parts among different models, the value of `model_type` must not be hardcoded. The preferred solution is to take the information about the model type from the namelist `master_model_nml` by calling `get_my_process_name()`.

**Open Question:**\
As the argument `model_type` of `vlr_add` turned out to be mandatory in order to ensure a working restart, there is the open question whether `model_type` should become a mandatory, rather than optional, subroutine argument?

@u241211

Approved-by: Florian Prill <florian.prill@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>