[nwp] Introduce Merge Request template
See merge request icon/icon-nwp!1322

## Feature Description
Add a MR template for nwp-features. This should be the base of all future MR.
It should describe the most important steps of the review process and guide users through the
ICON jungle.
Special emphasis is put keeping the template informative, but not overloading it with too much content.
One could add more template for bugfixes or other organizations.

This template is inspired by the on from https://gitlab.dkrz.de/icon-libraries/libfortran-support, where bugfix and feature
templates have proven to be very useful for both developer and maintainer.

This description already has the structure of the template.

Approved-by: Daniel Hupp <daniel.hupp@meteoswiss.ch>
Approved-by: Guenther Zaengl <guenther.zaengl@dwd.de>
Approved-by: Xavier Lapillonne <xavier.lapillonne@meteoswiss.ch>
Approved-by: Ralf Mueller <ralf.mueller@dkrz.de>
Merged-by: Daniel Hupp <daniel.hupp@meteoswiss.ch>