[nwp] 1mom-microphysics granule: Build and Testing Infrastructure
See merge request icon/icon-nwp!1251

# Build and Testing for 1-moment microphysics granule
A set of instruction about this new level of testing can be found in https://gitlab.dkrz.de/icon/wiki/-/wikis/Granule-Testing
## Why this MR
This infrastructure introduces enhanced testing for ICON, providing an additional layer of evaluation. It can be referred to as unit tests, as it allows the examination of specific components, like microphysics independently. For that an independent build and test infrastructure has been established on purpose.

## Structure
Introduce a new folder in root `test/microphysics_1mom_schemes` that contains all extra infrastructure to 
build, run and test the granules.
This separation of folder prevents any bad interference with the icon build-system for regular build of ICON.

## Build
The granule-test is built using CMake. The recipe is located in `test/microphysics_1mom_schemes/CMakeList.txt`.   

## Run
The driver is `test/microphysics_1mom_schemes/microphysics_1mom_driver.f90`.
It reads in an input-data set, runs a scheme and write the output back to file.
The driver is GPU-capable and kept as minimal as possible.
Therefore it is only able to run the full set of testcases for each granule, but not individual ones.

## Test
For each granule a meaningful number of test has been selected.
They basically cover all runtime-options passable as argument to the subroutines. These include:
- compute t-tendencies
- compute q-tendencies
- assimilate lhn
- verbosity (extra debug prints)

Probtest is used to generate and check a set of references and tolerances as we do for BuildBot.
The tolerances are machine- and compiler-agnostic, therefore stored in the repo itself.

Users are not required to generate new tolerances on their own. Gitlab-Ci will generate new references and upload
them as artifact on request. 

## CI-Integration
The granules is tested using gitlab-ci. Same for the generation of new tolerances and references. 
The runner offering levante sw-stack is used.
Compilation and testing requires ~2 minutes.

This additional level of testing is in place for Intel and GCC.
More setups and compilers could come into place, once the beta-phase of this testing is over.

Merged-by: Daniel Hupp <daniel.hupp@meteoswiss.ch>