[nwp] nwp_ocean_coupling: correct calculation of Qtop, re-enable LSM patching
See merge request icon/icon-nwp!1281

Qtop is the actual melt potential, the flux designated for melting
ice / snow. The ocean uses it as such regardless of temperature
(temperature of the ice is not even tracked in the ocean). Previously,
Qtop was passed from the NWP physics package as the net radiation at
the top of the ice sheet. This lead to ice melting whenever there was
downward net flux, regardless of temperature of the ice. This
changeset takes the melt potential directly from the sea-ice scheme.
Like the heat flux at the bottom, `condhf_ice`, it is provided as a
field `meltpot_ice` that can be written to file. To better diagnose
atmosphere-ocean coupling issues, the NWP-VDIFF interface also
provides water vapor fluxes for each surface type, as
`flx_water_vapor_{land,{l,s}{wtr,ice}}`. These are given in
kg/(m**2 s), positive downward.

Also re-enables patching of the LSM for runs with VDIFF/JSBACH. The
decision to patch is made opportunistically: if the extpar file
contains an ocean LSM in the form of a field `cell_sea_land_mask`,
the patch is performed. If not, we assume that the file is compatible
with the ocean LSM already, e.g., because it was modified externally.
A log message is issued in both cases so users can confirm that the
simulation runs off the intended LSM.

Approved-by: Stephan Lorenz <stephan.lorenz@mpimet.mpg.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>