Bugfix in modis cdnc
See merge request icon/icon-nwp!1389

Modis Cdnc was read in cloud_num in mo_nwp_phy_init only at the beginning of a simulation. It was fine when we did monthly simulations. But in recent yearly simulations, we saw that the cloud_num is the same throughout the months.

With the fix, cndc is updated once per day together with the daily update of other external parameter fields.

Approved-by: Axel Seifert <axel.seifert@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>