[nwp] Only attempt to read latbc file if latbc option in tracers.xml is set to 'file'
See merge request icon/icon-nwp!1385

This MR avoids any attempt to read ART AEROSOL tracers data from the lateral BC file when the `latbc` option in the _tracers.xml_ file is not set to `file`.

The original issue was that MeteoSwiss was not able to correctly run their operational cycle with pollen (ART) active, because of a call to finish in ICON

```plaintext
FINISH PE: 1 mo_async_latbc_utils::fetch_from_buffer_3D_generic_pollbetu: Internal error, invalid field index! Is pollbetu listed in latbc dict?
```

that was being triggered by the fact that the `latbc` option for pollen tracers was empty (i.e., not `file`), but the lateral BC file was still containing the data relative to pollen, which in turn, due to how this is implemented in ICON, would allow an unnecessary/unwanted attempt to read the pollen variables (i.e., setting `lread_tracer` to true). However, this would result in the error above because the following function was not able to find the requested variable:

```plaintext
!-------------------------------------------------------------------------
!>
! Return the index for a given variable in mapped variable list
!
FUNCTION get_field_index(buffer,name) RESULT(result_varID)
  TYPE(t_buffer), INTENT(IN) :: buffer
  CHARACTER (LEN=*),   INTENT(IN) :: name !< variable name
  ! local variables
  LOGICAL, PARAMETER :: ldebug = .FALSE.
  INTEGER :: result_varID, varID, nvars
  CHARACTER(len=len(name)) :: name_lcresult_varID = -1
  nvars = buffer%ngrp_vars
  if (nvars >= 1) name_lc = tolower(name)
  ! looping over variable list in internal name
  LOOP : DO varID=1, nvars
    IF (name_lc == tolower(buffer%internal_name(varID))) THEN
      result_varID = varID
      EXIT LOOP
    END IF
  END DO LOOP
END FUNCTION get_field_index
```

Note that the variable could not be found because `buffer%ngrp_vars`, which depends on the number of variables in group `LATBC_PREFETCH_VARS`, was not set to a value large enough to reach the `varID` corresponding to the pollen tracers in the loop above. Finally, the reason why the pollen variables are not part of `LATBC_PREFETCH_VARS` is that pollen tracers are only made part of that group if the `latbc` option in the _tracers.xml_ file is set to `file`, which brings us back to beginning of this explanation.

So, in short MeteoSwiss would like to be able to keep the pollen data in the lateral BC file, but (only) change the _tracers.xml_ file to control whether the data must be read or not. However, the current situation is that ICON would not throw an error only in the following cases:

* The lateral BC file does not contain pollen data and the `latbc` option for the respective tracer in the _tracers.xml_ file is set to empty.
* The lateral BC file contains pollen data and the `latbc` option for the respective tracer in the _tracers.xml_ file is set to `file`.

However, the two things (data present in the lateral BC file and `latbc` option in _tracers.xml_ set to `file`) are treated separately in ICON, leading to these kind of (inconsistency) errors. Therefore, this MR allows (only for ART AERSOL tracers) to control both directly from the _tracers.xml_ option, making sure that there is no attempt to read the variable if the _XML_ entry does not specify that.

To do this we set `lread_tracer` to false if the corresponding ART AEROSOL tracer is not part of the `LATBC_PREFETCH_VARS` group.

_For more information ask_ @b382090. _The issue is a bit complex to explain with text, but I will do my best to answer questions if there are any._

Approved-by: Andrea Leuthard <andrea.leuthard@meteoswiss.ch>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>