CAMS forecasted aerosol as an option for ecRad radiation
See merge request icon/icon-nwp!1180

The 3D CAMS forecasted aerosols are introduced as an input to radiation. The CAMS forecast is available twice a day in a 3 hours temporal resolution. The data is horizontally interpolated using CDO (as done with CAMS climatology) as a pre-processing step. The vertical interpolation is done inside ICON. There are currently 11 aerosols species and the data is of mixing ratios (kg/kg). The interactions of these aerosols with radiation are already implemented in ecRAD (including the optical properties). The reading of the file is done in the same way as for CAMS climatology with slight changes.

For running this option the user should set irad_aero = 8 in the Namelist.

The second Namelist parameter is the path to the file containing the CAMS forecast: cams_aero_filename. This parameter is identical for both CAMS climatology and CAMS forecast (irad_aero = 7, 8).

In this merge request we also made a change in the DeMott ice nucleation parametrization scheme. In the previous version one could choose to couple the DeMott formula with CAMS/Tegen aerosols (icpl_aero_ice = 1 or 2) regardless of the aerosols-radiation interaction choice (irad_aero). In this version there is consistency between the two schemes. When DeMott scheme is chosen by icpl_aero_ice = 1 (instead of 0 for the default Cooper formula) it is automatically coupled with the aerosols that were chosen for radiation (irad_aero = 6 for Tegen or 7 for CAMS climatology). Therefore, icpl_aero_ice = 2 is no more valid.

Approved-by: Martin Köhler <martin.koehler@dwd.de>
Merged-by: Ulrich Schaettler <ulrich.schaettler@dwd.de>