Tuning parameters for TERRA-URB
See merge request icon/icon-nwp!1400

Tuning parameters are introduced for the anthropogenic heat flux (AHF) and the impervious surface area (ISA) in order to improve the specifcation of these fields without the need of detailed external parameter data. The new AHF specification depends on the climatological 2m-temperature rather than being a globally constant value, and the ISA is parameterized based on a spatially smoothed urban fraction, assining higher ISA to city centres than to suburbian regions. A more detailed description is given in the namelist doc.

Approved-by: Jan-Peter Schulz <jan-peter.schulz@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>