[mpim] Introduce two-step compilation (separate module files)
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/428

This introduces the '--enable-separate-mods' configure option. When the option
is enabled, the compilation of Fortran source files is split into two steps: the
module file generation and the object file generation. Although this increases
the total overhead (especially the filesystem part of it) of the compilation, it
has the potential to reduce the wall clock time of the parallel builds.

Merged-by: Sergey Kosukhin <sergey.kosukhin@mpimet.mpg.de>