[nwp] nwp_sfc_interface: set runoff on lakes to precipitation minus evaporation
See merge request icon/icon-nwp!1366

Currently, precipitation falling on lakes leaves the model, while
evaporation from lakes is created from nothing. This is a leak in
the model's water balance.

Without modeling the water content of lakes explicitly, we cannot
physically fix this leak. But we can account for the water entering
or leaving the model and at least close the global water balance.

This changeset sets the surface runoff on lakes to the difference
between precipitation and evaporation. This difference would result
in a change of the lake water level, and by assigning it to runoff,
we effetively keep the lake water level constant. The runoff can be
passed to an HD model to close the global water balance. Locally,
the continuity equation d/dt (total water content) = P - E - Runoff
now holds on all non-ocean tiles.

NOTICE: the reference area of all non-tile runoff variables changes
from "per square meter of land" to "per square meter of grid box".

The drawback is that lakes may produce negative runoffs when there
is more evaporation than precipitation. Those reduce the grid-scale
runoff, simulating the feeding of the lake from surrounding land
tiles, and they reduce the outflow of the HD model, simulating
rivers replenishing the lake water level.

Approved-by: Dmitrii Mironov <dmitrii.mironov@dwd.de>
Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>