[nwp] Removed the alteration of Modis Cdnc
See merge request icon/icon-nwp!1422

We want to include this change into a 'release' branch for icon-xpp (previously called icon-seamless). Since this is very small change, it would be good if we can get this merged back quickly. 

In previous version, we changed the values of Modis Cdnc in certain area, mainly increasing them. This alteration was hard-coded in ICON. With new tests, we decided that without the data alteration, the results are better. Also in the future, if there is any correction in the satellite data necessary, I recommend that should be done out of ICON code.

Approved-by: Axel Seifert <axel.seifert@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>