[mpim] ICON-Land: fixes for using older restart files
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/433

The recently merged pond scheme (!375) in JSBACH introduced several new variables
that are written (and therefore required) in the restart file. Some of them are not
needed in the restart no matter if the pond scheme is active or not, others are only
needed if the pond scheme is active.

This creates a problem if one wants to use a restart file from an experiment before
the merge of the pond scheme even if the pond scheme is not active.
This is fixed in this MR.

See https://gitlab.dkrz.de/jsbach/jsbach/-/merge_requests/185

Merged-by: Reiner Schnur <reiner.schnur@mpimet.mpg.de>