[nwp] fix in call of MPI aerosol routines
See merge request icon/icon-nwp!1363

In the psrad interface, aerosol routines were called with jcs=1, but blocks with jcs>1 were shifted.
As this shifting is not done in the ecrad interface, the MPI aerosol routines have to be called with jcs=jcs.

Additionally, it was found in MR !1263  that with irad_aero=18, experiments with a limited-area domain or nest did not pass nproma tests.
Therefore, irad_aero=18 was activated in clmexp.ICON_CLM and the icon-dev checksuite was added to enable the nproma test (and the restart and mpi tests, see checksuite.clm/check.ICON_CLM).

Approved-by: Roland Wirth <roland.wirth@dwd.de>
Merged-by: Daniel Rieger <daniel.rieger@dwd.de>