[nwp] CDNC scaling for climate projections
See merge request icon/icon-nwp!1420

To make the external CDNC climatology consistent with the climate scenarios, a scaling factor for the cdnc's (scaling `cloud_num` in the code) is introduced. It can be switched on by setting `lscale_cdnc = .true.` , but it can only be used in combination with the simple plume scheme and the MODIS cdnc's (`icpl_aero_gscp = 3`). If one of these is not set, it is reset to `.false.`
The factor can be written out as `cloud_num_fac`. In the current version, the scaling is limited to 0.1 - 3 (however, the largest value, even for the year 2099 and SSP585, is 1.15, and the lowest one for SSP119 is 0.77).

Concerning the implementation, the factor is equal to `x_cdnc` from the simple plume scheme (https://gmd.copernicus.org/articles/10/433/2017/, equ. 15) scaled with its value for the year 2005. The `x_cdnc` alone would not work as it was designed as a factor on a background cdnc, but using the MODIS cdnc we already have a realistic field for which only the anthropogenic part is influenced by any climate scenario.

Approved-by: Trang Van Pham <trang.pham-van@dwd.de>
Merged-by: Daniel Rieger <daniel.rieger@dwd.de>