[mpim] ICON-Land: improvements in JSBACH soil hydrology
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/402

This ICON-Land update includes extensive changes to the JSBACH soil hydrology. It introduces
a new soil property, the residual water content, as lower limit for vertical water transport
within the soil column. This branch also replaces the soil field capacity with porosity as
upper limit for maximum soil water content - wherever this was still used in the soil physics.
However, it does not yet touch the use of field capacity as upper limit for root zone moisture
and plant water stress.

Furthermore, infrastructure was implemented to choose between a semi-distributed hydrology
parameterization that accounts for sub-grid variability in infiltration and subsurface drainage
(the Arno scheme, to be used for coarse-scale simulations), and an alternative parameterization
that assumes a uniform distribution of soil moisture. The latter is still experimental and should
(later) be used for high resolution, site-level or HydroTiles applications.

Finally, a few bugfixes are included.

This commit changes results.

Further details are found in the corresponding ICON-Land MR
https://gitlab.dkrz.de/jsbach/jsbach/-/merge_requests/166.

The impacts of these revision are documented in issue
https://gitlab.dkrz.de/jsbach/jsbach/-/issues/209 as experiment `SoilHydrRev`.

Co-authored-by: Philipp De-Vrese <philipp.de-vrese@mpimet.mpg.de>
Co-authored-by: Reiner Schnur <reiner.schnur@mpimet.mpg.de>
Co-authored-by: Veronika Gayler <veronika.gayler@mpimet.mpg.de>
Merged-by: Reiner Schnur <reiner.schnur@mpimet.mpg.de>