[hotfix] Increase sanitizer timeout for GPU test
See merge request icon/icon-nwp!1442

Cherry pick from hotfix in icon-main.

https://gitlab.dkrz.de/icon/icon/-/merge_requests/387

In case of file system maintenance at CSCS the Buildbot working directory for
the machine Balfrin is moved from /scratch to /opr. In such cases, the
compute-sanitizer used for some GPU tests takes a very long time to attach to
the icon processes, which causes the test to fail. This hotfix simply increases
the timeout to let the test run successfully.

Approved-by: Xavier Lapillonne <xavier.lapillonne@meteoswiss.ch>
Merged-by: Ulrich Schaettler <ulrich.schaettler@dwd.de>