[nwp] bugfix: improve robustness of horeka config wrappers
See merge request icon/icon-nwp!1438

`config/kit/hk.cpu.gnu-13-openmpi-5.0` did not compile when `cuda.h` is found in CPATH. This is the case for the default Horeka user environment, but not the environment that the buildbot service account uses. This MR improves the robustness of the Horeka config wrappers with respect to some environment variables.

Approved-by: Michael Krayer <michael.krayer@dwd.de>
Merged-by: Ulrich Schaettler <ulrich.schaettler@dwd.de>