[mpim] Adjust nproma to 256 B alignment to improve GPU performance
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/444

The performance on AMD GPUs may suffer significantly (5-10% for whole model, up
to 15-20% for radiation) if nproma / nproma_sub is not aligned to 256 bytes. On
NVIDIA GPUs (at least A100), the effect is more modest, but 256-byte alignment
seems to improve performance a little bit (1-2%) also there.

This MR makes the proper adjustments when using nblocks_c=1 and nblocks_sub=x.

Merged-by: Sergey Kosukhin <sergey.kosukhin@mpimet.mpg.de>