Initialise aerosol fields in case of Kinne/CAMS Aerosol
See merge request icon/icon-nwp!1446

The fields zaeq_* are not used and therefore not intialised in case of Kinne/CAMS aerosol. This creates however error in BBtest. 

https://bb-prod8.dkrz.de/#/builders/51/builds/4616

https://bb-prod8.dkrz.de/#/builders/51/builds/4616/steps/4/logs/checksuite_clm_check_ICON_CLM

This MR adds a hotfix for the initialisation of the fields (just in time for the current OSR). A proper solution needs to be added at a later stage (i.e. only in case of reduced radiation, avoid code duplication etc.)

Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Approved-by: Daniel Rieger <daniel.rieger@dwd.de>
Merged-by: Ulrich Schaettler <ulrich.schaettler@dwd.de>